requirements.txt
# E-Commerce MVP - Unified Requirements
# 单体架构依赖（预留微服务迁移接口）

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0  # Async SQLite for future scaling

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2  # 直接使用 bcrypt，避免 passlib 兼容性问题
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Image Processing
pillow==10.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0  # For future microservice communication

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Production (预留)
# gunicorn==21.2.0
# redis==5.0.1  # For future caching/session
# celery==5.3.6  # For future async tasks
